requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
google-cloud-vision==3.4.4
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0